+++ src/tiled/scriptdialog.cpp
@@ -31,17 +31,17 @@
 #include <QComboBox>
 #include <QCoreApplication>
 #include <QDoubleSpinBox>
+#include <QGroupBox>
 #include <QHBoxLayout>
 #include <QJSEngine>
 #include <QLineEdit>
 #include <QPixmap>
 #include <QPushButton>
+#include <QRadioButton>
 #include <QSet>
 #include <QSize>
 #include <QSlider>
 #include <QTextEdit>
-
-#include <memory>
 
 static const int leftColumnStretch = 0;
 // stretch as much as we can so that the left column looks as close to zero width as possible when there is no content
@@ -104,6 +104,30 @@ public:
     Q_INVOKABLE void addItems(const QStringList &texts)
     { QComboBox::addItems(texts); }
 };
+
+
+void ScriptButtonGroup::addItems(const QStringList &values, const QStringList &toolTips)
+{
+    int toolTipIndex = 0;
+    for (const QString &value : values) {
+        addItem(value, toolTips.value(toolTipIndex));
+        toolTipIndex++;
+    }
+}
+
+QAbstractButton *ScriptButtonGroup::addItem(const QString &value, const QString &toolTip)
+{
+    QRadioButton *radioButton = new QRadioButton(mLayout->parentWidget());
+    radioButton->setText(value);
+    if (!toolTip.isEmpty())
+        radioButton->setToolTip(toolTip);
+
+    mLayout->addWidget(radioButton);
+
+    QButtonGroup::addButton(radioButton, QButtonGroup::buttons().length());
+
+    return radioButton;
+}
 
 
 ScriptDialog::ScriptDialog(const QString &title)
@@ -267,6 +291,19 @@ ScriptDialog::NewRowMode ScriptDialog::newRowMode() const
     return m_newRowMode;
 }
 
+ScriptButtonGroup *ScriptDialog::addRadioButtonGroup(const QString &labelText,
+                                                     const QStringList &values,
+                                                     const QString &toolTip,
+                                                     const QStringList &buttonToolTips)
+{
+    QGroupBox *groupParent = new QGroupBox(this);
+    QHBoxLayout *hBox = new QHBoxLayout(groupParent);
+    ScriptButtonGroup *buttonGroup = new ScriptButtonGroup(groupParent, hBox);
+    buttonGroup->addItems(values, buttonToolTips);
+    addDialogWidget(groupParent, labelText, toolTip);
+    return buttonGroup;
+}
+
 void ScriptDialog::setNewRowMode(NewRowMode mode)
 {
     m_newRowMode = mode;
@@ -278,7 +315,9 @@ int ScriptDialog::exec()
     return QDialog::exec();
 }
 
-QWidget *ScriptDialog::addDialogWidget(QWidget *widget, const QString &label)
+QWidget *ScriptDialog::addDialogWidget(QWidget *widget,
+                                       const QString &label,
+                                       const QString &labelToolTip)
 {
     determineWidgetGrouping(widget);
     if (m_widgetsInRow == 0)
@@ -292,6 +331,8 @@ QWidget *ScriptDialog::addDialogWidget(QWidget *widget, const QString &label)
 
     if (!label.isEmpty()) {
         QLabel *widgetLabel = newLabel(label);
+        if (!labelToolTip.isEmpty())
+            widgetLabel->setToolTip(labelToolTip);
         widgetLabel->setSizePolicy(QSizePolicy::Minimum, QSizePolicy::Fixed);
         widgetLabel->setBuddy(widget);
         m_rowLayout->addWidget(widgetLabel);